requirements.txt
pandas
numpy
joblib
scikit-learn==1.1.2
Flask
tensorflow-cpu
gunicorn